requirements.txt
streamlit>=1.38,<2
qrcode==7.4.2
pillow>=11.0.0
